+++ R/center.R
@@ -69,3 +69,54 @@ geos_centerish <- function(shp, epsg = 3857) {
   cent
 }
 
+#' Get the centroid of the maximum inscribed circle
+#'
+#' Returns the centroid of the largest inscribed circle for each shape
+#'
+#' @param shp An sf dataframe
+#' @templateVar tolerance TRUE
+#' @templateVar epsg TRUE
+#' @template template
+#'
+#' @return An sf dataframe where geometry is the circle center of each shape in shp
+#' @export
+#'
+#' @concept center
+#'
+#' @examples
+#' data(towns)
+#' st_centerish(towns)
+#' 
+st_circle_center <- function(shp, tolerance = 0.01, epsg = 3857) {
+  
+  cent <- geos_circle_center(shp, tolerance = tolerance, epsg = epsg)
+  
+  sf::st_geometry(shp) <- sf::st_as_sfc(cent)
+  shp
+}
+
+#' Get the centroid of the maximum inscribed circle
+#'
+#' Returns the centroid of the largest inscribed circle for each shape
+#'
+#' @param shp An sf dataframe
+#' @templateVar tolerance TRUE
+#' @templateVar epsg TRUE
+#' @template template
+#'
+#' @return A geos geometry list
+#' @export
+#'
+#' @concept center
+#'
+#' @examples
+#' data(towns)
+#' geos_centerish(towns)
+#' 
+geos_circle_center <- function(shp, tolerance = 0.01, epsg = 3857) {
+  
+  shp <- make_planar_pair(x = shp, epsg = epsg)$x
+  
+  geos::geos_centroid(geos::geos_maximum_inscribed_crc(shp))
+}
+