+++ man-roxygen/template.R
@@ -1,1 +1,2 @@
 #' <%=ifelse(exists("epsg") && epsg, "@param epsg numeric EPSG code to planarize to. Default is 3857.", "") %>
+#' <%=ifelse(exists("tolerance") && tolerance, "@param tolerance postive numeric tolerance to simplify by. Default is 0.01.", "") %>