+++ R/geomander-package.R
@@ -14,6 +14,7 @@ NULL
 #' @importFrom tidyr pivot_wider
 #' @importFrom stringr str_detect
 #' @importFrom readr read_csv write_csv
+#' @importFrom rlang .env .data :=
 NULL
 
 globalVariables(c('.'))