+++ NEWS.md
@@ -1,5 +1,6 @@
 # geomander 2.1.0
 * Avoid indirect download for `get_alarm()`. Fixes bug where block level data resulted in mismatched geometry.
+* Add `st_circle_center()` and `geos_circle_center()` for centroid of the maximum inscribed circle.
 
 # geomander 2.0.2
 * Drop `spdep` dependency to avoid timeouts on checking examples.